requirements.txt
autoflake==1.2
beautifulsoup4==4.6.3
black==18.3a0
datasketch==1.2.10
diskcache==3.1.1
gspread==3.0.1
gspread_dataframe==3.0.1
isort==4.3.4
oauth2client==4.1.3
google-api-python-client==1.7.4
flake8==3.6.0
mypy==0.641
nltk==3.3
pandas==0.23.4
psycopg2-binary==2.7.5
profilehooks==1.10.0
snowflake-connector-python==1.6.12
yapf==0.25.0
seaborn==0.9.0
